+++ tests/testthat/test-merging.R
@@ -5,24 +5,25 @@ context("Merging titers")
 
 # Setup merge tests
 titer_merge_tests <- tibble::tribble(
-  ~titers,                      ~conservative,  ~likelihood,  ~lispmds,
-  c("*"),                       "*",            "*",          "*",    # * unchanged
-  c("10"),                      "10",           "10",         "10",   # numeric unchanged
-  c("<10"),                     "<10",          "<10",        "<10",  # less than unchanged
-  c(">80"),                     ">80",          ">80",        ">80",  # greater than unchanged
-  c("."),                       ".",            ".",          ".",    # . unchanged
-  c(10, 10),                    "10",           "10",         "10",   # two same numeric, merges same
-  c("*", "10", "10", "*"),      "10",           "10",         "10",   # two same numeric, two *,  merges same
-  c("<10", "<40", "<20", "*"),  "<10",          "<10",        "<10",  # all less than go to min
-  c(">10", ">40", ">20", "*"),  ">40",          ">40",        ">40",  # all greater than to to max
-  c("10", "160", "2560", "*"),  "160",          "160",        "*",    # very variable not set to . (default sd_limit=NA), including *
-  c("<10", "20", "*", "*"),     "<40",          "10",         "<40",  # mix of numeric & less than
-  c("<10", "20", "*", "."),     "<40",          "10",         "<40",  # mix of numeric & less than
-  c("*", "*", "*", "*"),        "*",            "*",          "*",    # all * merge to *
-  c(".", "."),                  ".",            ".",          ".",    # all . merge to .
-  c(".", "*"),                  "*",            "*",          "*",    # mix of . & * merge to *
-  c("<20", "20", ">20"),        "*",            "*",          "*",    # mix of numeric, less than, and greater than, becomes *
-  c("<20", "40", ">10", "."),   "*",            "*",          "*",    # mix of numeric, less than, greater than and ., becomes *
+  ~titers,                      ~conservative,  ~likelihood,  ~lispmds,  ~conservativesd1, ~likelihoodsd1,
+  c("*"),                       "*",            "*",          "*",       "*",              "*",             # * unchanged
+  c("10"),                      "10",           "10",         "10",      "10",             "10",            # numeric unchanged
+  c("<10"),                     "<10",          "<10",        "<10",     "<10",            "<10",           # less than unchanged
+  c(">80"),                     ">80",          ">80",        ">80",     ">80",            ">80",           # greater than unchanged
+  c("."),                       ".",            ".",          ".",       ".",              ".",             # . unchanged
+  c(10, 10),                    "10",           "10",         "10",      "10",             "10",            # two same numeric, merges same
+  c("*", "10", "10", "*"),      "10",           "10",         "10",      "10",             "10",            # two same numeric, two *,  merges same
+  c("<10", "<40", "<20", "*"),  "<10",          "<10",        "<10",     "<10",            "<10",           # all less than go to min
+  c(">10", ">40", ">20", "*"),  ">40",          ">40",        ">40",     ">40",            ">40",           # all greater than to to max
+  c("10", "160", "2560", "*"),  "160",          "160",        "*",       "*",              "*",             # very variable not set to . (default sd_limit=NA), including *
+  c("<10", "20", "*", "*"),     "<40",          "10",         "<40",     "*",              "*",             # mix of numeric & less than
+  c("<10", "20", "*", "."),     "<40",          "10",         "<40",     "*",              "*",             # mix of numeric & less than
+  c("*", "*", "*", "*"),        "*",            "*",          "*",       "*",              "*",             # all * merge to *
+  c(".", "."),                  ".",            ".",          ".",       ".",              ".",             # all . merge to .
+  c(".", "*"),                  "*",            "*",          "*",       "*",              "*",             # mix of . & * merge to *
+  c("<20", "20", ">20"),        "*",            "*",          "*",       "*",              "*",             # mix of numeric, less than, and greater than, becomes *
+  c("<20", "40", ">10", "."),   "*",            "*",          "*",       "*",              "*",             # mix of numeric, less than, greater than and ., becomes *
+  c("<10", "320"),              "<640",         "40",         "*",       "*",              "*",             # variable mix of < and numeric
 )
 
 test_that("Test log titer conversion", {
@@ -61,6 +62,18 @@ test_that("Test titer merging", {
     expect_equal(
       ac_merge_titers(titer_merge_tests$titers[[x]], options = RacMerge.options(method = "lispmds")),
       titer_merge_tests$lispmds[x]
+    )
+
+    # Conservative result with sd limit
+    expect_equal(
+      ac_merge_titers(titer_merge_tests$titers[[x]], options = RacMerge.options(method = "conservative", sd_limit = 1)),
+      titer_merge_tests$conservativesd1[x]
+    )
+
+    # Likelihood result with sd limit
+    expect_equal(
+      ac_merge_titers(titer_merge_tests$titers[[x]], options = RacMerge.options(method = "likelihood", sd_limit = 1)),
+      titer_merge_tests$likelihoodsd1[x]
     )
 
   }
@@ -128,7 +141,7 @@ test_that("Merge sd_lim working", {
   )
 
   expect_equal(
-    "<40",
+    "*",
     ac_merge_titers(c("<10", "20"), options = RacMerge.options(sd_limit = 0.9))
   )