+++ NEWS.md
@@ -189,3 +189,4 @@
 # Racmacs 1.1.40
 * Add some options to view.acmap to set the starting viewer translation, rotation and zoom
 * Add `grid.lwd` option to control grid linewidth in `ggplot(map)`
+* Add antigen and sera names to vector when retreiving antigen and sera attributes like `agFill()`