+++ NEWS.md
@@ -196,4 +196,5 @@
 * Add new options and default methods to use when merging titers as described for `RacMerge.options()`
 * Return a more useful error if `bootstrapBlobs()` is run on a map where `bootstrapMap()` has not yet been performed.
 * Add `grid.margin.lwd` option to control grid margin linewidth in `ggplot(map)`
+* Improve user interruption of map optimization runs