+++ NEWS.md
@@ -197,3 +197,5 @@
 * Add `grid.margin.lwd` option to control grid margin linewidth in `ggplot(map)`
 * Improve user interruption of map optimization runs
 
+# Racmacs 1.2.1
+* Titer merge method changed so that sd limit is implemented before any merging to < values